+++ src/states/common/mod.rs
@@ -6,10 +6,9 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-mod approved;
 mod base;
 
-pub use self::{approved::Approved, base::Base};
+pub use self::base::Base;
 use crate::time::Duration;
 
 /// Delay after which a bounced message is resent.
